Support local variables in methods.
SOM has static scopes, which makes the compiler's job easy. This commit is
sufficient for programs such as the following to compile and run:

  Hello = (
    run = (
      | t1 t2 |
      t1 := 'Hello '.
      t2 := t1 concatenate: 'world'.
      t2 println.
    )
  )

printing, as expected, "Hello world".

+++ src/lib/compiler/instrs.rs
@@ -13,8 +13,8 @@ pub enum Instr {
     Primitive(Primitive),
     Return,
     Send(usize),
-    VarLookup(u32),
-    VarSet(u32),
+    VarLookup(usize),
+    VarSet(usize),
 }
 
 #[derive(Clone, Copy, Debug)]
@@ -27,4 +27,4 @@ pub enum Primitive {
 }
 
 /// The index of the "self" variable.
-pub const SELF_VAR: u32 = 0;
+pub const SELF_VAR: usize = 0;
